apps/soliso/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
psycopg2-binary
python-multipart
python-jose[cryptography]
pillow
aiofiles